API/requirements.txt
annotated-types==0.7.0
blessed==1.20.0
boto3==1.34.123
botocore==1.34.123
certifi==2024.6.2
cffi==1.16.0
chalice==1.31.2
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
idna==3.7
inquirer==2.10.1
jmespath==1.0.1
jwt==1.3.1
marshmallow==3.21.3
packaging==24.1
pipdeptree==2.22.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
PyYAML==6.0.1
readchar==4.1.0
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
typing_extensions==4.12.2
urllib3==1.26.18
wcwidth==0.2.13
chalice
geocoding
psycopg2-binary
python-dotenv